Fix dataset pickling (#7)
* Add test to trigger pickling error

Computing a dataset using "processes" or a Client scheduler currently
fails due to a pickling error. This adds a corresponding test that will
need to be fixed.

* Fix failing pickling test

`PosixPath.open()` returns a `BufferedReader` that is not pickleable,
breaking dask computation with `Client` or "processes" schedulers. This
still uses `resources` to identify data file paths, but opens directly
from the paths instead of file buffers to fix pickling.