Refactor from single dispatched methods to inherited methods
The single dispatching system became a little convoluted with the
introduction of the estimator wrapper, and required duplication of
things like the Dataset and DataArray methods. I switched to an
inheritance system where each image type has an ImageWrapper subclass
that 1) defines the preprocessor for the image type and 2) implements
the sklearn methods.