Estimator wrapper (#18)
* Implement estimator wrapper with kneighbors and predict

* Wrapper tests

* Improve type annotations

* Simplify is_image_type check

* Fix single-output prediction by storing fitted metadata and auto-squeezing y

* Ignore coverage for type checking blocks

* Test var names are set correctly

* Make utils module

* Make wrapped method signatures consistent and improve type hints

* Docstrings

* Refactor from single dispatched methods to inherited methods

The single dispatching system became a little convoluted with the
introduction of the estimator wrapper, and required duplication of
things like the Dataset and DataArray methods. I switched to an
inheritance system where each image type has an ImageWrapper subclass
that 1) defines the preprocessor for the image type and 2) implements
the sklearn methods.

* Hopefully fix type annotation...

* Almost definitely fix type annotations

* Address review comments

Better naming and type annotations, remove redundant fitted checks,
test for expected NotFittedErrors.