Improve ambiguous type annotations (#22)
* Overload load_swo_ecoplot with correct return types

* WIP decorators pass annotations through (mypy not passing)

* Get decorator passthrough passing type checks

- Ditched the type aliases that were causing the TypeVar parsing
  to fail.
- Made the decorated AttrWrapper generic

* Overload kneighbors return type

* Set default for overloaded load_swo_ecoplot

* Add typing-extensions for PEP 612 support in 3.9

We now rely on the ParamSpec and Concatenate features of PEP 612
which weren't implemented until Python 3.10. In order to use those
features while supporting Python 3.9, we can temporarily replace
any typing imports with backported features from typing-extensions.
Once we drop 3.9 support, we can switch back to typing and drop the
dependency.

* Fix default type signature for load_swo_ecoplot

To infer the correct type when `as_dataset` is unspecified,
we shouldn't give it a default value in the `True` case.

* Include n_neighbors and return_distance in all kneighbors signatures

* Fix DaskBacked kneighbors shape with n_neighbors

The number of output targets for kneighbors must be n_neighbors when
it's specified.