Change default output_sizes to number of bands instead of 1
It seems reasonable to make the default output consistent with the
simplest possible use case of an identity function lambda x: x. In
other words, the default case will be a function that returns one
band for each input band, rather than collapsing all the input bands
to a single output.