Add image ufunc API (#28)
* WIP quick and dirty ufunc prototype implementation

Tests are completely broken and a lot of features are missing, but
this is a basic working prototype of the core features that should
allow for some preliminary performance comparisons.

* WIP set gufunc params based on estimator metadata

* WIP rough implementation of image subclasses

* WIP restore dimension order after applying ufunc

* Re-implement feature name checks

* Remove unnecessary type check

* Fix: kneighbors now returns dimension k=1 instead of variable=k1

* keep_attrs=True to preserve CRS

* Reorganize tests and get all passing

* Change default output_sizes to number of bands instead of 1

It seems reasonable to make the default output consistent with the
simplest possible use case of an identity function lambda x: x. In
other words, the default case will be a function that returns one
band for each input band, rather than collapsing all the input bands
to a single output.

* Fix outdated comment

* Let numpy infer nodata dtype

* Remove default values for ufunc output params

Because there's no "typical" way to use ufuncs, coming up with
unsurprising defaults probably isn't possible. For now, we'll
settle for no defaults, but we may consider trying to infer
outputs based on a small test run, similar to what
dask.array.apply_gufunc does with output dtypes.