Fix NaN masking by filling on-the-fly instead of at init
I also moved nan_fill from an init parameter to an apply parameter,
since it's more relevant to the function being applied than an
attribute of the image chunk.

I compared performance of filling with np.where versus copying
and filling in-place by indexing, and the former was marginally
faster.