WIP Only allocate mask array and cast flat array if needed
I had to add a bunch of NDArray casts to make mypy happy, since it
can't follow the logic that `ImageChunk.nodata_mask` is not None
at runtime in the cases where it's being used. It would be better
to avoid that union type for `nodata_mask`, but I'm not sure how to
do that without needlessly allocating an unused array yet.

quickfix