Disable nan_fill by default for apply_ufunc_across_bands
It was already disabled by default for UfuncArrayProcessor.apply,
so this makes them consistent. Defaulting to 0 would be safer for
functions that fail on NaN inputs, but modifying input data is more
surprising.

Also, fix test_nan_filled that wasn't correctly using the parametrized
value. Now the test ensures that the nan_fill parameter is set in the
func input, instead of just checking that np.nan isn't present. In
that context, parametrizing over skip_nodata doesn't make sense, so
I removed that.