Temporarily use int32 minimum as kneighbours nodata_output
Since we can only assign a single nodata_output, we need a default
value that works for both distances and indexes. Negative numbers
should be safe in most cases since distances and array indexes will
always be positive. The neighbor ID data type is system-dependent,
so this should work for either 32-bit or 64-bit systems.