Implement NoData skipping and other `apply` params (#41)
* WIP first stab at nodata skipping (tests failing, no tuple support)

Quick prototype implementation of an option for skipping nodata
pixels in applied functions rather than filling and masking them.
Predicting from a heavily masked numpy array with an RF model is
2.5x faster with skipping vs filling.

* WIP add unmasked pixel hack for fully masked arrays

* WIP refactoring

* WIP re-add support for tuple outputs like kneighbors

* WIP Only allocate mask array and cast flat array if needed

I had to add a bunch of NDArray casts to make mypy happy, since it
can't follow the logic that `ImageChunk.nodata_mask` is not None
at runtime in the cases where it's being used. It would be better
to avoid that union type for `nodata_mask`, but I'm not sure how to
do that without needlessly allocating an unused array yet.

quickfix

* WIP Fix missing arg and use of local nan_fill

* WIP remove  and add  params (tests failing)

* Passing tests

* More tests

* Allow selecting image types when parametrizing

* Store nodata_output in _FillValue attr

* Refactor out repeated nodata_output check

* Switch from `prevent_empty_array` to adjustable `ensure_min_samples`

* _FillValue won't be set on Dataset vars by default

* Rename _ImageChunk to UfuncArrayProcessor and split into ufunc module

Also expanded docstring and reorganized methods to approximate
execution order, plus some minor unrelated refactoring.

* Type the ufunc callable

* Add decorators for mapping functions and methods over tuples

* Last minute renaming

* Add `allow_cast` param to cast output dtype to fit nodata_output

* Import Concatenate from typing_extensions for 3.9

* Typing fix for 3.9

* Use nan_fill for ensuring min samples

* Cast output to min dtype for integers

Casting arrays directly to the nodata_output type would use
int64 for any integer. Using np.min_scalar_type ensures we get the
minimum type for ints. Using that for floats would cast int arrays to
float16, which seems wrong, so currently I use the nodata_output
type in those cases, which for floats gives float64 arrays.

* Add optional check for `nodata_output` in ufunc output

* Fix bug where valid pixels are overwritten to ensure min samples

Previously, ensuring 3 min samples would always set 3 dummy values
at the start of the flat array, but this could overwrite valid data.
Instead, we need to only set dummy values in masked pixels, which
can be done by indexing with the nodata mask.

This also refactors to avoid the repeated cast of nodata_mask.

* Mention that allow_cast prevents error in nodata_output docstring

* Fix UfuncArrayProcessor docstring

* Refer to NaN instead of np.nan in docstring

We should generally refer to the concept of NaN unless it's specific
to the numpy implementation of NaN.

* One-line docstring formatting

* Order methods chronologically

* Refer to samples instead of pixels

If we separate the flattening of images from UfuncArrayProcessor,
it makes more sense to refer to samples rather than pixels
throughout.

* Use defined nodata_input

* Disable nan_fill by default for apply_ufunc_across_bands

It was already disabled by default for UfuncArrayProcessor.apply,
so this makes them consistent. Defaulting to 0 would be safer for
functions that fail on NaN inputs, but modifying input data is more
surprising.

Also, fix test_nan_filled that wasn't correctly using the parametrized
value. Now the test ensures that the nan_fill parameter is set in the
func input, instead of just checking that np.nan isn't present. In
that context, parametrizing over skip_nodata doesn't make sense, so
I removed that.

* Temporarily use int32 minimum as kneighbours nodata_output

Since we can only assign a single nodata_output, we need a default
value that works for both distances and indexes. Negative numbers
should be safe in most cases since distances and array indexes will
always be positive. The neighbor ID data type is system-dependent,
so this should work for either 32-bit or 64-bit systems.

* Fix missing nan_fill with predict and kneighbors

This previously defaulted to 0.0, but now needs to be set
explicitly.