Support wrapped methods with n-dimensional data (#50)
* WIP partial support for 1D arrays

So far this is mostly just addressing dimensions more generally
so they're not explicitly 2D, which seems to get us most of the
way to 1D support. We were failing 1 test as we require 1D samples
to store features as the first dimension, which is consistent with
numpy images and xarray images and samples, while Pandas-like
samples are stored with features as the last dimension. For now,
that's fixed by simply transposing the inputs and outputs.

* Fix test by transposing 1D samples

* Test predict with 2-5 dimensions

* Fix ndim comparison

* Explain why 1D inputs/outputs are transposed