Fix overload order for Pylance
Pylance seems to show the first defined overload as the default,
so it should match the actual default value for the function to
avoid confusion. This wasn't previously the case for kneighbors or
load_swo_ecoplot.

Overloads don't seem to be very clearly specified, so this may not
work for every language server. See:
https://discuss.python.org/t/some-improvements-to-function-overloads/36275/3