Support multiple `nodata_output` values (#58)
* Implement map_over_arguments

* Validate mapped argument names

* Support multiple nodata_output values for kneighbors

* Set nodata_output for kneighbors dynamically depending on return_distance

I mistakenly thought that the overloads were changing the defaults, but
those only affect the language server, not the actual runtime. The
nodata_output is now set dynamically based on return_distance, and a
test was added to confirm that works as expected.

* Fix overload order for Pylance

Pylance seems to show the first defined overload as the default,
so it should match the actual default value for the function to
avoid confusion. This wasn't previously the case for kneighbors or
load_swo_ecoplot.

Overloads don't seem to be very clearly specified, so this may not
work for every language server. See:
https://discuss.python.org/t/some-improvements-to-function-overloads/36275/3

* Allow mapping over positional arguments

Extracting mapped arguments only from kwargs required them to be
passed by name, which was very easy to accidentally forget, leading
to confusing type errors. By using inspect.signature.bind, we can
identify mapped arguments positionally as well.

* Raise on multiple nodata values when return_distance is False