Fix inverse_transform feature name bug
When an estimator is fit without feature names (i.e. with a numpy
array), it stores an empty array of feature names in the wrapped
meta. That correctly allows for feature name warnings when predicting
with a dataframe, but causes an issue when applying inverse_transform
to a named feature array, e.g. an xr.Dataset as it attempts to
assign those empty feature names to the Dataset.

To solve that, we need to 1) store the number of features seen during
fitting, and 2) use sequential integers as the feature names when
none are available. This is what we already do with unnamed targets.