Test attrs for all wrapped methods
This is a pretty inelegant way to test the methods, but it should
be temporary until we get regression tests in place that will test
attrs implicitly.

The test did reveal a bug where old attrs were overriding new attrs
when keep_attrs=True, so this fixes that as well.