Drop inherited attrs and set metadata attrs on estimator results (#65)
* Reset attrs and set long_name

* Set source_method attr

* Test dropping and setting attrs

* List test dependencies in CI

Helpful for debugging test failures. Note that we need to premake
the environment in order to list the dependencies before running
the tests.

* Avoid drop_attrs bug in xarrary<2025.03.0

Using drop_attrs(deep=False) doesn't work in older xarray versions,
but we can achieve the same thing by simply resetting attrs. See
https://github.com/pydata/xarray/issues/10027

* Refactor getting source_method attr

* Append ufunc method name to history attr

* Drop _FillValue from global attrs

* Add keep_attrs param to all estimator methods

* Test attrs for all wrapped methods

This is a pretty inelegant way to test the methods, but it should
be temporary until we get regression tests in place that will test
attrs implicitly.

The test did reveal a bug where old attrs were overriding new attrs
when keep_attrs=True, so this fixes that as well.

* fix typo

* Clarify wording