Provide generic `bool` overload for `synthesize_feature_array`
mypy requires a fallback overload to allow passing a generic
`bool` to a `Literal[True] | Literal[False]` parameter. We're
not currently doing that, but this is consistent with the
overloads for `_generate_fractal_noise` and avoids the potential
surprise in the future.

See https://mypy.readthedocs.io/en/stable/literal_types.html