Make test config global and patch Pooch for releases
After bumping the package version but before releasing the corresponding
tag, Pooch would fail to fetch data that's required for testing because
it attempts to download from the nonexistent tag. This fixes that by 
conditionally patching the fetcher to an unversioned one that downloads 
from main if the current package version isn't found in the git tags. 

For that patch fixture to apply to doctests, I needed to move conftest.py
out of the tests directory and into the root directory so that it would
be in scope for the src directory. That constraint isn't well documented
in pytest, but running `pytest --fixtures src/` prior to the fix shows
that the patch fixture was unavailable to the doctests.