Treat NaN as a valid NoData value
I was considering None and NaN to both indicate missing NoData,
but that isn't necessary and forces _get_nodata_mask to use the
slower path with an extra comparison any time that NaN is present.
Instead, it should be treated like any other NoData value and will
correctly warn and be ignored if it is specified but doesn't fit 
in the array dtype.