Avoid `object` dtype `nodata_input` by using masked arrays (#93)
* Use a MaskedArray to represent nodata_input

This allows us to reliably identify missing 
NoData values (i.e. features that should not be
masked). Previously, we used None which would
create object type arrays that have terrible
performance.

* Optimize a fast path when all NoData are valid

* Warn and skip NoData values that don't fit the feature array dtype

* Use consistent types when passing NoData values to _build_masked_nodata_array

* Treat NaN as a valid NoData value

I was considering None and NaN to both indicate missing NoData,
but that isn't necessary and forces _get_nodata_mask to use the
slower path with an extra comparison any time that NaN is present.
Instead, it should be treated like any other NoData value and will
correctly warn and be ignored if it is specified but doesn't fit 
in the array dtype.

* Fix indexing bug with uncastable values

Enumerating over just the non-None values caused the index
to be wrong. Now we enumerate over all the NoData values and
just skip Nones.

* Explicitly set 0 as the masked fill value

We're not actually using the masked array fill value, but we
can specify it for consistency in case it's used in the 
future.

* Use nodata_input.data directly to avoid the overloaded operator creating a temporary masked array

* Fix outdated comments

* Error instead of warning for uncastable nodata_input

* Fix typo and reword comment